Add some components to FireWeatherPage
such as PageHeader, PageTitle, and WeatherStationDropdown
and fix broken tests
